allow assignment types to be edited and deleted

+++ resources/views/gradebook/assignment_grade.blade.php
@@ -138,7 +138,7 @@
                             if (row.points_earned === null || row.is_excused) {
                                 return '--';
                             }
-                            $number = parseInt(row.points_earned) / parseInt(row.assignment.max_points) * 100;
+                            $number = parseFloat(row.points_earned) / parseInt(row.assignment.max_points) * 100;
                             return $number + '%';
                         }
                     },
